clean premier_res and def_func
