+++ utilities/def_func.py
@@ -19,6 +19,8 @@ import glob
 from typing import Union
 import sys
 import bisect
+from astral.sun import sun
+import astral
 
 def get_detection_files(num_files: int) -> List[str]:
     """Opens a file dialog multiple times to get X APLOSE formatted detection files.
@@ -864,7 +866,47 @@ def input_date(msg, tz_data):
     
     return date_dt
 
+def suntime_hour(begin_deploy, end_deploy, timeZ, lat,lon):
+    """ Fetch sunrise and sunset hours for dates between date_beg and date_end
+    Parameters :
+        date_beg : str Date in format 'YYYY-mm-dd'. Start date of when to fetch sun hour
+        date_end : str Date in format 'YYYY-mm-dd'. End date of when to fetch sun hour
+        timeZ : tz_data, FixedOffset object of pytz module
+        lat : str latitude in Decimal Degrees
+        lon : str longitude in Decimal Degrees
+    Returns :
+        hour_sunrise : list of float with sunrise decimal hours for each day between date_beg and date_end 
+        hour_sunset : list of float with sunset decimal hours for each day between date_beg and date_end 
+    """    
+    # Infos sur la localisation
+    gps = astral.LocationInfo( timezone=timeZ,latitude=lat, longitude=lon)
+    # List of days during when the data were recorded
+    list_time = pd.date_range(begin_deploy, end_deploy)
+    h_sunrise = []
+    h_sunset = []
+    dt_dusk = []
+    dt_dawn = []
+    astral.Depression = 12 # nautical twilight see def here : https://www.timeanddate.com/astronomy/nautical-twilight.html
+    # For each day : find time of sunset, sun rise, begin dawn and dusk
+    for day in list_time:
 
+        suntime = sun(gps.observer,date=day, dawn_dusk_depression = astral.Depression)
+        # suntime = sun(gps.observer,date=day)
+        dawn_dt=(suntime['dawn'])
+        
+        dusk_dt=(suntime['dusk'])
+        
+        day_dt=(suntime['sunrise'])
+        
+        night_dt=(suntime['sunset'])
+        
+        day_hour = day_dt.hour+day_dt.minute/60
+        night_hour = night_dt.hour+night_dt.minute/60
+        h_sunrise.append(day_hour)
+        h_sunset.append(night_hour)
+        dt_dusk.append(dusk_dt)
+        dt_dawn.append(dawn_dt)
+    return h_sunrise, h_sunset, dt_dusk, dt_dawn
